+++ source/Core/Utils.php
@@ -1353,7 +1353,7 @@ class Utils extends \OxidEsales\Eshop\Core\Base
         $oView = oxNew(\OxidEsales\Eshop\Application\Controller\FrontendController::class);
         $oView->init();
         $oView->render();
-        $oView->setClassName('oxUBase');
+        $oView->setClassKey('oxUBase');
         $oView->addTplParam('sUrl', $sUrl);
         if ($sRet = \OxidEsales\Eshop\Core\Registry::getUtilsView()->getTemplateOutput('message/err_404.tpl', $oView)) {
             $sReturn = $sRet;